+++ selfdrive/ui/qt/offroad/offroad_home.cc
@@ -1,102 +1,7 @@
-#include "selfdrive/ui/qt/home/home.h"
+#include "selfdrive/ui/qt/offroad/offroad_home.h"
 
-#include <QHBoxLayout>
-#include <QMouseEvent>
-#include <QStackedWidget>
 #include <QVBoxLayout>
-
-#include "selfdrive/ui/qt/offroad/experimental_mode.h"
-#include "selfdrive/ui/qt/util.h"
-#include "selfdrive/ui/qt/widgets/prime.h"
-
-// HomeWindow: the container for the offroad and onroad UIs
-
-HomeWindow::HomeWindow(QWidget* parent) : QWidget(parent) {
-  QHBoxLayout *main_layout = new QHBoxLayout(this);
-  main_layout->setMargin(0);
-  main_layout->setSpacing(0);
-
-  sidebar = new Sidebar(this);
-  main_layout->addWidget(sidebar);
-  QObject::connect(sidebar, &Sidebar::openSettings, this, &HomeWindow::openSettings);
-
-  slayout = new QStackedLayout();
-  main_layout->addLayout(slayout);
-
-  home = new OffroadHome(this);
-  QObject::connect(home, &OffroadHome::openSettings, this, &HomeWindow::openSettings);
-  slayout->addWidget(home);
-
-  onroad = new OnroadWindow(this);
-  slayout->addWidget(onroad);
-
-  body = new BodyWindow(this);
-  slayout->addWidget(body);
-
-  driver_view = new DriverViewWindow(this);
-  connect(driver_view, &DriverViewWindow::done, [=] {
-    showDriverView(false);
-  });
-  slayout->addWidget(driver_view);
-  setAttribute(Qt::WA_NoSystemBackground);
-  QObject::connect(uiState(), &UIState::uiUpdate, this, &HomeWindow::updateState);
-  QObject::connect(uiState(), &UIState::offroadTransition, this, &HomeWindow::offroadTransition);
-  QObject::connect(uiState(), &UIState::offroadTransition, sidebar, &Sidebar::offroadTransition);
-}
-
-void HomeWindow::showSidebar(bool show) {
-  sidebar->setVisible(show);
-}
-
-void HomeWindow::updateState(const UIState &s) {
-  const SubMaster &sm = *(s.sm);
-
-  // switch to the generic robot UI
-  if (onroad->isVisible() && !body->isEnabled() && sm["carParams"].getCarParams().getNotCar()) {
-    body->setEnabled(true);
-    slayout->setCurrentWidget(body);
-  }
-}
-
-void HomeWindow::offroadTransition(bool offroad) {
-  body->setEnabled(false);
-  sidebar->setVisible(offroad);
-  if (offroad) {
-    slayout->setCurrentWidget(home);
-  } else {
-    slayout->setCurrentWidget(onroad);
-  }
-}
-
-void HomeWindow::showDriverView(bool show) {
-  if (show) {
-    emit closeSettings();
-    slayout->setCurrentWidget(driver_view);
-  } else {
-    slayout->setCurrentWidget(home);
-  }
-  sidebar->setVisible(show == false);
-}
-
-void HomeWindow::mousePressEvent(QMouseEvent* e) {
-  // Handle sidebar collapsing
-  if ((onroad->isVisible() || body->isVisible()) && (!sidebar->isVisible() || e->x() > sidebar->width())) {
-    sidebar->setVisible(!sidebar->isVisible());
-  }
-}
-
-void HomeWindow::mouseDoubleClickEvent(QMouseEvent* e) {
-  HomeWindow::mousePressEvent(e);
-  const SubMaster &sm = *(uiState()->sm);
-  if (sm["carParams"].getCarParams().getNotCar()) {
-    if (onroad->isVisible()) {
-      slayout->setCurrentWidget(body);
-    } else if (body->isVisible()) {
-      slayout->setCurrentWidget(onroad);
-    }
-    showSidebar(false);
-  }
-}
+#include <QStackedWidget>
 
 // OffroadHome: the offroad home page
 
@@ -241,3 +146,4 @@ void OffroadHome::refresh() {
     alert_notif->setText(QString::number(alerts) + (alerts > 1 ? tr(" ALERTS") : tr(" ALERT")));
   }
 }
+
